locator: implement ReplicasOrdered (ring-wise)
In order to be consistent with LWT optimisation implementation in other
drivers, it is mandatory to return query plan consisting of replicas
put in ring order. To accomplish that, the (abstractly unordered)
ReplicaSet becomes convertible into a newly defined ReplicasOrdered
struct. Iterating over ReplicasOrdered yields replicas in ring order, so
policies can depend on it to perform LWT optimisation correctly.

Note that the computation of ReplicasOrdered is lazy and performed in
two steps.
- the first step is relatively cheap and free of allocations
and involves finding the primary replica by iterating over the global
ring. It is triggered by the first call to ReplicasOrderedIterator::next().
- the second step is relatively expensive and involves allocations.
It computes the whole ReplicaArray of all replicas, put in the ring
over.
Precomputation is not an option, as it would be very expensive
to compute and store all precomputed arrays that would respect the
global ring order (as opposed to only local, per-DC ring order).