default_policy: harden ReplicaLocationCriteria
It is cleaner when the Criteria itself carries the required information
for the filtering, as well as less prone to errors (because one has to
provide the required DC/rack name).
As a bonus, the semantics of helper fns calls in `pick()` were made more
sane: `pick_replica()` is called with
`ReplicaLocationCriteria::Datacenter` only if the preferred datacenter
is actually specified. `fallback()` was modified analogously (in regard
to calling `shuffled_replicas()`).