connection: make it possible to switch off write coalescing
Each connection is managed by a tokio task that reads from and writes to
a single socket. The task (called internally as "router") polls, amongst
others, a future that collects write requests from a channel and pushes
them to the socket. It currently has a heuristic that is meant to be a
throughput optimization: if there are no more write requests on the
channel, the future yields, hoping to give an opportunity for more write
requests to arrive on the channel. This improves efficiency as it allows
more requests to be batched together into a single write and reduces the
number of syscalls issued.

The described optimization was implemented in:
e1c0b6f3b48eac2f60c043c05fc0b806189788c4

However, this optimization obviously can cause worse latency in some
scenarios. For example, if there aren't many requests being issued but
the application otherwise runs lots of unrelated tokio tasks, then a
yielded task can be rescheduled again only after a significant delay. If
the request rate is low enough then batching won't ever bring any
benefit but will only make the latency worse.

In order to help in those scenarios, this commit adds a flag that allows
turning off the write coalescing optimization. Write coalescing is kept
turned on by default, but it can be easily disabled when constructing a
new session.

Results of a test with modified cql-stress that runs 100 tasks which
spin for 0.25ms and then yield, in a loop
(https://github.com/piodul/cql-stress/commit/f20050e3a3514af20f431e0c4c187cd254500fbd):

  cargo run --release --bin cql-stress-scylla-bench -- -max-rate 10000
      -mode write -workload uniform -partition-count 128 -concurrency 16
      -nodes 127.0.0.1 -duration 10s
      -noise-task-count 100 -noise-task-hog-duration 250us

With the coalescing optimization _enabled_:

  99h:          11.5ms
  95h:          7.21ms
  90h:          6.25ms
  median:       3.74ms
  mean:         4.11ms

With the coalescing optimization _disabled_:

  99h:          6.24ms
  95h:          4.89ms
  90h:          4.25ms
  median:       2.32ms
  mean:         2.44ms