session_test: add a SimpleStrategy test
Scylla is moving away from SimpleStrategy,
in general users should use NetworkTopologyStrategy everywhere now.

Still it's good to test that the driver works with a SimpleStrategy keyspace,
it's still a case that is supported, even if it's slowly getting deprecated.

This test runs a few queries in a SimpleStrategy keyspace.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>