examples: remove clone.rs example
clone.rs is an old example that uses SessionBuilder
to build many copies of Session.

I think this example is harmful because it promotes
the use of multiple Sessions, which is usually a bad idea.

One user started using multiple Sessions instead of Arc<Session>
because they saw such approach in this example:
https://github.com/scylladb/scylla-rust-driver/issues/724#issuecomment-1557645457

The example itself isn't very useful. It just uses
a Builder to Build sessions, this isn't anything tricky,
it's what's expected of the Builder pattern.

I think it would be better to remove it to avoid
misleading the users.

Fixes: #727

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>