exec_profile: Default to SerialConsistency::LocalSerial
Contrary to what is stated in documentation, the default
SerialConsistency was None, which resulted in SerialConsistency not
being set in the CQL frames. This led to ProtocolError in case of LWT
statements. As it is better to always include SerialConsistency (even
for statements for which it is not relevant), the previous default is
reverted.