latency_awareness: Add `scale` parameter
The `scale` parameter was accidentally omited when porting
LatencyAwareness implementation from other drivers.
Its purpose is to control the weight that the lastly observed latency
has in the newly computed average (relative to the previous average).
To the best of my understanding of the previous version of the code,
the scale that was implicitly used before was equal to **1 second**.
The new default value is, in line with other drivers, **100
milliseconds**.