integration: add tests for [serial] consistency
One test checks for each [serial] possible consistency that the request
frame indeed contains the specified [serial] consistency.
Another test checks that RoutingInfo's consistencies match the specified
ones. As the latter test confirmed the bug in Session regarding
RoutingInfo, it is temporarily marked as ignored. The mark will be
removed in the next commit, after the bug is fixed.