session: fail only if no hostnames can be resolved
It makes more sense to fail only when we end up with no contact points
with known resolved addresses. In that case, a new variant is returned:
NewSessionError::FailedToResolveAnyHostname, else initialisation
continues successfully.
However, to warn the user that some hostnames could not be resolved,
a warn message is issued.