session: Impl calculate_token_for_partition_key()
So far, the driver has provided a way to calculate token based on
a PreparedStatement and values for that statement. However, if
a user knows:
  - the order of the columns in the partition key,
  - the values of the columns of the partition key,
  - the partitioner of the table that the statement operates on,

then having a PreparedStatement is not necessary and the token can be
calculated based on that information. The function
calculate_token_for_partition_key() is added, which does that
calculation, along with a test.