partitioner: Transition to stateful partitioners
The Partitioner API is changed into stateful one, resembling std::hash
traits. The motivation is simple: having Partitioners hold state enables
feeding them with data in chunks instead of one contiguous slice, which
saves us an allocation.
Since now, Partitioners' users are encouraged to first build a hasher
using `Partitioner::build_hasher()` and then feed it in chunks with
`write()`. Finally, `finish()` is to be called and a token is returned.
The `Partitioner::hash()` method was renamed to `hash_one()` for closer
correspondence to `std::hash::BuildHasher` trait.