CI/Makefile: use "--wait" in "docker compose up"
When using "--wait" flag (introduced in Docker Compose V2) in 
"docker compose up", Docker waits for all healthchecks to be in 
"healthy" state. Since docker-compose.yml files in this repository
use "cqlsh" to determine the health of a container, this means
that Docker will wait for all nodes of the cluster to be fully
bootstrapped and their CQL port ready.

By using this flag, this fixes a potential problem in CI - previously
the code would not wait for the last node to be fully started. 
Additionally, "wait.sh" script can be deleted (this logic now handled by
Docker Compose).