routing: remove Node struct
Remove struct scylla::routing::Node from routing.rs. This structure
was introduced in 2020 in 1c8df890d6419e4612fbdbb4aef1fb68c35ffb88
and it is unused in our current code (and it looks like the original
commit didn't even use it at all). 

Moreover, this struct clashes with scylla::transport::Node making it
error-prone to find the wrong Node struct for a user of the driver. 
The deleted Node structure also relied solely on SocketAddr, while we 
want to identify the nodes by host_id (as in scylla::transport::Node).

Refs #660.