connection: remove re-export of QueryResult
Remove re-export of QueryResult (from scylla::transport::query_result::QueryResult
to scylla::transport::connection::QueryResult). The motivation for this
change is to reduce the number of re-exports in the codebase (some of
them don't make sense and are there for legacy reasons), as a part
of our effort of stabilizing the API.

It probably makes sense to follow this up with moving QueryResult
from transport/ module to a more appropriate module (query/ module?).

Refs #660