frame: reduce visibility of several funcs/consts
Reduce visibility of several functions and constants which are only
used internally within frame and don't have to be exposed publicly.
In this patch only the most obvious ones are tackled (ones that require
a change solely in scylla-cql). This change is a part of our effort
to stabilize the API.

Refs #660