session: remove estimate_replicas_for_query
Remove estimate_replicas_for_query from Session. A user of this function
can switch to using get_token_endpoints() from ClusterData directly.

It did not make sense for estimate_replicas_for_query() to be in
Session, as this is too much low-level function for Session and doesn't
fit the abstraction of Session.

As a followup, it might make sense to look into whether 
get_token_endpoints() from ClusterData should get removed as well:
ReplicaLocator can be used directly.

This change is a part of our effort to stabilize the API and reduce
the number of pub functions.

Refs #660