topology: refactor to prepare for fallback
The code of `read_metadata()` is refactored to prepare for addition of
fallback logic.
A new log is added at debug level for a successful metadata fetch.
Sadly, the new design requires cloning peers due to the borrow checker.
Luckily, this only happens if the first fetch attempt fails and is not
particularly expensive.