topology: add fallback logic to read_metadata()
This is the final step in boosting the driver's robustness in case of
sudden IP changes in the cluster. In case that the control connection
fails to fetch metadata, after all known peers fail to be queried, the
fallback phase is entered: addresses of all initially known nodes are
resolved again and fetching metadata from them is attempted.

Until now, if all nodes in the cluster changed their IPs at once (which
is particularly easy in a single-node cluster), the driver was unable to
contact the cluster ever more. Now, the driver reconnects the control
connection with the next metadata fetch attempt (after at most 60
seconds, as configured in ClusterWorker), fetches new topology,
discovers all nodes' addresses and finally connects to them.

The next commits are yet to shorten this downtime from 60 seconds to
possibly 1 second, by reacting immediately to the control connection
being broken.