prepared: introduce PartitionKey struct
It is clear that the algorithm for computing partition key for a given
prepared statement and bound values can be divided into steps. The first
of them involves extracting values that consistute the partition key and
putting them in proper partition key order. PartitionKey struct performs
this step on construction, and serves as an entry point for further
steps, with token calculation among others. For those further steps to
be done, an iterator over partition key values is accessible, which
yields pairs (value, spec) for each column that constitutes partition
key, in partition key order. It will be used to avoid materialising
partition key in a buffer.