topology: unpub Metadata
`Metadata` serves the mere purpose of keeping the metadata together
until it gets incorporated into `ClusterData`. It it at no point
visible to the user by any public interface, so we can unpub it.
Disclaimer: this is a breaking change, as the type was reachable from
outside. I can't think of a useful way of utilising it, though, so
nobody should be hurt.