default_policy: Prefer local rack nodes
Until now, the default policy respected rack preference only when it
came to replicas. Considering non-token-aware queries (e.g. unprepared
statements), they would never prioritize the preferred rack.
Now, local rack nodes are always preferred before other nodes, both in
`pick()` and in `fallback()`. Tests and docs are adjusted for this
behaviour. One new test is added for the case when no token is provided
and a rack is preferred.