session: require schema agreement to fetch schema version
It can be a nasty source of race bugs when schema version is fetched
without awaiting for schema agreement. It is especially risky when one
awaits schema agreement and subsequently fetches schema version, because
the schema version can change in the meantime and schema may no longer
be in agreement.
To cope with the problem, `await_schema_agreement()` now returns the
agreed schema version, and `fetch_schema_version()` is removed
alltogether.