cluster: elide clones by moving keyspaces there and back
Keyspace strategies were cloned and allocated into a Vec, just for
creating an iterator to references to them in order to construct a
`ReplicaLocator`. It was done due to lifetime issues (in order to send
references to another thread, they must be 'static).
The problem is solved by sending keyspaces to another thread for
constructing `ReplicaLocator` and then moving it back to the calling
thread afterwards. Clones are elided. Hooray!