session: await_schema_agreement() gets implicit timeout
Similarly to
- what Java driver 3 does in regard to awaiting schema
agreement's API,
- what was recently done in Rust driver in regard to tracing info's API,

the API of awaiting schema agreement is altered so that:
- `await_timed_schema_agreement()`, which takes explicit timeout, is
  removed,
- `await_schema_agreement()` is now bound with a timeout that is set
  globally per-`Session` in `SessionConfig`.

The motivation is that it could lead to application's deadlock when
`await_schema_agreement()` was called with no timeout given and the
cluster stoped synchronising (so schema agreement would be never
reached), It is rarely desirable to block the calling application for
arbitrarily long time, especially that some extreme situations are
possible, such as a network partition between nodes.

`SessionBuilder`s API is accomodated to the changes. Docs are updated as
well.