proxy: make workers aware of the target shard
It can be useful to have proxy workers aware of the shard they are
connected to. This is done the same way as Scylla itself does it, that
is, by taking the driver's port modulo shards count. This, of course,
requires the proxy to know the shards count, which depends on set
ShardAwareness and the cluster's shard awareness.