session: no materialising partition key for display
Traces generated upon requests are expected to contain the partition key
if available. In order to avoid an allocation for that, `PartitionKey`
struct is used, whose iterator yields the required values one by one
and then they are deserialised.
Partition key is extracted only once, and token is calculated only once,
too.

Not to raise our MSRV, `Option::unzip()` is temporarily reimplemented
as an util `unzip_option()`. It is to be replaced with `Option::unzip()`
once MSRV is raised to at least 1.66.