dev-deps: downgrade criterion to 0.4
The `criterion` library, since version 0.5, depends on `clap` v4 or
higher. Recently, `clap` 4.4.0 has been released which has bumped its
required MSRV to 1.70.0. Our MSRV is still 1.65.0. Apparently, the cargo
resolver isn't smart enough to choose an older version of `clap` which
still supports 1.65.0, so this breaks our MSRV build in CI
(see rust-lang/cargo#9930).

In order to unblock the CI without checking in Cargo.toml - which would
be annoying - the `criterion` dependency is downgraded to 0.4. This
library is only used for benchmarks and is a dev-dependency, so it's not
a big deal. No changes in the code are necessary.

In the future, if we decide to bump MSRV to 1.70.0 or further, we can
update the dependency.