treewide: fix clippy complaints about useless use of vec!
Clippy started to suggest using arrays instead of `vec!` where
applicable, i.e. if the vector would be converted to an iterator anyway,
then we can get rid of an allocation by replacing it with an array. This
is only an issue with a small number of tests and it's easy to fix, so
this commit just applies clippy's suggestions.