test_coalescing: make sure that query response is not an error
`test_coalescing` runs a bunch of insert queries using `Connection::query`.
It unwraps the result of `Connection::query`, but this function returns
a `QueryResponse`, which might be `QueryResponse::Error`.
To make sure that the query was successful we have to check that the response
doesn't contain any errors. This can by done using `into_non_error_query_response()`.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>