cassandra/docker-compose: go back to using the latest cassandra version
In January the tests started failing on Cassandra `4.1`, so we pinned the cassandra
version to `4.0.7`, which didn't have these problems. `4.0.7` is the version that we've
been using in the CI since then.

In the meantime there were new releases of `Cassandra`, at the moment the latest is `4.1.3`.
Let's try to go back to the latest version, maybe the problems fixed themselves during the last nine months.

Refs: https://github.com/scylladb/scylla-rust-driver/issues/620
Refs: https://github.com/scylladb/scylla-rust-driver/pull/629

Fixes: https://github.com/scylladb/scylla-rust-driver/issues/633

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>