benches: make them compile again
The recent fix that changed the `types::write_short` function to use u16
instead of i16 broke one of our benchmarks - it tried to call that
function with a negative number. The benchmarks aren't compiled in our
CI, so it didn't catch it.

Fix the issue by using `u16::MAX` instead of `-1`.