scylla-macros: implement enforce_order flavor of SerializeCql
Some users might not need the additional robustness of `SerializeCql`
that comes from sorting the fields before serializing, as they are used
to the current behavior of `Value` and properly set the order of the
fields in their Rust struct. In order to give them some performance
boost, add an additional mode to `SerializeCql` called "enforce_order"
which expects that the order of the fields in the struct is kept in
sync with the DB definition of the UDT.

It's still safe to use because, as the struct fields are serialized,
their names are compared with the fields in the UDT definition order
and serialization fails if the field name on some position is
mismatched.