serialize/value: make fallbacks for `SerializeCql` explicit
Currently, there is a blanket implementation of `SerializeCql` for all
types that implement `Value`. Replace it with an `impl SerializeCql`
block for every type that currently implement `Value`.

This is just a preparation - the blocks that were introduced still
utilize the same fallback logic as the blanket impl. They will be
replaced with proper impls in the following commits.