frame/value_tests: exercise both old and new ser cql traits
Adjust the existing tests for the `Value` trait to also do serialization
via `SerializeCql` and compare results produced by both traits.

The tests pass at the moment because the new trait is implemented using
the old one. In the commits that follow, we will introduce proper
implementations of the new trait for the existing types. In nearly all
cases we will be interested in preserving compatibility, so the exising
tests will be very helpful.