frame/value: don't require trait on MaybeUnset parameter
Currently, the `MaybeUnset` wrapper requires the inner type to implement
`Value`. As we want to use the type parameter both with the old and the
new serialization framework, this restriction doesn't make much sense in
the case of the new framework, so remove it. The impls for `Value` and
`SerializeCql` on the `MaybeUnset` itself should themselves require
appropriate traits to be implemented on the inner type.