serialize/row: make fallbacks for `SerializeRow` explicit
Like it was done for `SerializeCql`, remove the blanket `ValueList` ->
`SerializeRow` impl and replace with explicit `impl` blocks for all
types for which we currently implement `ValueList`.

The newly introduced `impl SerializeRow` blocks use the same fallback
logic as the original blanket implementation - it will be replaced in
further commits.