connection.rs: Don't allow queries with values
After serialization refactor it will be impossible to perform unprepared
query with values, because serializing values will require knowing
column types.

In order to make refactor easier and better split responsibility, this
commit removes `values` arguments from `Connection` methods, so that
it is callers responsibility to prepare the query if necessary.