frame/value: introduce ValueOverflow error
Some conversions between time-related types return `ValueTooBig` in case
when one representation doesn't fit into the other. This is a
misinterpretation of the `ValueTooBig`'s semantics: it is supposed to be
returned when the value's size in bytes is longer than the maximum size
of `i32::MAX`, not when the value overflows the allowed range (e.g.
doesn't fit in a 64bit number).

Introduce `ValueOverflow` error in order to differentiate the cases when
a value is "too big" and use it in conversions between time-related
types.

Adjust relevant `Value` implementations which use the conversions to map
the `ValueOverflow` to `ValueTooBig`. This is done in order not to
introduce breaking changes to the soon-to-be-legacy interface.