serialize/writers: make overflow errors easier to handle
The CQL format imposes the following limits:

- At most u16::MAX values to be provided to a query,
- At most i32::MAX bytes in a single value.

Currently, the interface defined in the types::serialize::writers module
checks those limits and panics if they are exceeded during
serialization.

While those limits are quite large, they sometimes can accidentally be
exceeded, and the legacy serialization code handles such situations
properly (i.e. it does not panic). We should not regress and make it
possible/easy to handle it in the new interface.

The following changes are made:

- `CellWriter::set_value` and `CellValueBuilder::finish` now return a
  Result, indicating whether there was an overflow or not.
- In order to account for the interface changes, `CountingWriter` is
  split into multiple separate types, for each of the traits that the
  original type implemented.
- The limit for u16::MAX values in `RowWriter` is lifted, and an `usize`
  is used to track the value count. The driver code is supposed to check
  whether the limit is exceeded or not before sending the query, and
  deferring the overflow check to that point should make the
  serialization code slightly more performant.