serialize/writers: add WrittenCellProof
The writer traits are using a type-level trick to force implementors of
`SerializeCql::serialize` to call `CellWriter::finish()`: the latter
returns `CellWriter::WrittenCellProof` which the former is required to
return. The fact that the proof is an associated type of the writer and
the `serialize` method is generic over the writer type makes sure that
the proof truly comes from the `finish()` call on the writer provided to
the method and not from any other writer. We are going to change
`CellWriter` into a struct, so this trick will no longer be applicable.

Instead, introduce a `WrittenCellProof` struct, which is a zero-sized
type generic over an invariant lifetime. The `CellWriter` struct will
consume itself and return a `WrittenCellProof` with the same lifetime
parameter as the original cell writer; `serialize` will require the
implementor to return it. The lifetime invariance makes sure that one
proof cannot be assigned to another if their lifetimes are not exactly
the same.