Public API: Use NewSerializedValues / SerializeRow
This is one of the main steps of serialization refactor: using new trait
in public API. Thanks to our implementations for this new trait
migrating should not be a huge issue for users - in many cases being
fully compatible.

One change missing from this commit is Batch support, as this requires
more changes (which will also be more breaking and require more
migration effort), so it will be implemented separately.