serialize/value: dedicated error variant for too big legacy values
Add a dedicated error variant to ValueToSerializeCqlAdapterError,
returned in case when the legacy implementation tried to serialize but
overflowed the maximum allowed size by the protocol. Previously, the
`ValueTooBig` error defined in frame::value would be returned, which was
inconsistent with other error cases for the `Value` -> `SerializeCql`
translation code.