serialize/value: fix error returned from dynamic tuples
In case when serialization of one of the fields fails, the tuple should
return an error indicating that serialization of the tuple failed, and
nest the error returned by field serialization inside it. The error used
to have the wrong ColumnType put into it - a loop variable containing
the ColumnType of the field shadowed the function argument containing
the ColumnType of the whole tuple. Fix the issue by renaming the loop
variable and un-shadowing the function argument as a result.