serialize/{row,value}: add tests for serialization errors
The PR that introduced impls of the SerializeRow/SerializeCql for the
types that used to implement the old traits relied on the tests in
`value_tests.rs` - they were extended with more test cases and modified
to run on both the old and the new traits. However, this was only done
for the tests that serialize stuff without causing errors. The new impls
return errors that are much more detailed than what the old ones used to
return, and - as evidenced by the bugs fixed in previous commits - badly
needed their own set of tests.

Add such a set. All different kinds of built-in errors should be
covered, with the exception of size overflow errors which would require
allocating huge amounts of memory to trigger them.