serialize/{row,value}: rename confusing "mismatch" variants
When serializing a Rust struct / CqlValue::UserDefinedType to a CQL UDT,
or a Rust struct / BTreeMap / HashMap as a list of named values, there
can be a fields/columns mismatch between the Rust and CQL definition
and a field/column can be present in one or missing in another.

Current error kind variants have very confusing or plainly wrong
names/descriptions/messages. Fix the situation in this commit by
enforcing a consistent naming scheme:

- If something from missing in Rust data but required by the CQL type,
  use `<RustStuff>MissingFor<CqlStuff>`.
- If something is present in Rust data but missing from the CQL type,
  use `NoSuch<CqlStuff>` or `No<CqlStuff>WithName`.

Apart from that, fix the docstrings and the error messages.