raw_batch: implement RawBatchValueAdapter
Implement an adapter layer which takes a `BatchValues` object (which
needs type information to serialize), pairs it with an iterator over
`RowSerializationContext` objects and returns something which implements
`RawBatchValues` (which don't require type information to serialize).

It will be used by the `scylla` crate to pass batch data to `scylla-cql`
in a type-erased form.