prepared_statement: add calculate_token_untyped method
The PreparedStatement::calculate_token method takes an object that
implements SerializeRow, serializes it to SerializedValues and returns
the token. For the sake of an optimization in the code that handles
batches we would like to provide a SerializedValues object directly.

Extract the part that calculates token from SerializedValues to a new,
separate method, but keep it pub(crate) - we'd like to avoid exposing
type-unsafe interfaces to the users.