statement/batch: reintroduce BatchValuesFirstSerialized
The purpose of the new struct is to enable token-aware routing of
batches - which already exists in the old API - in the new API.

Batches are routed according to the token calculated based on the first
statement in the batch (if the first statement is a prepared statement).
Calculation of the token must happen before the load balancing policy
computes a plan and chooses the first connection, but serialization only
happens after a connection is chosen. In order not to repeat
serialization work, BatchValuesFirstSerialized wrapper can be used to
transform a BatchValues into another BatchValues which caches the result
of the first serialization.

The types are put into a module in the `scylla` crate and hidden inside
it. The wrapping functionality is exposed via a function which
constructs the BatchValuesFirstSerialized object but returns it as `impl
BatchValues`.