scylla-macros: allow renaming fields in UDTs
In some cases, it might be desirable to name the Rust struct fields
differently than the UDT fields (e.g. due to a different naming
convention). Add an attribute to the `SerializeCql` macro, inspired by
serde's `rename` attribute, which causes the fields to be matched on a
specified name instead of the Rust name.