scylla-macros: allow renaming fields in rows
Like we just did for `SerializeCql`, introduce a `rename` annotation to
`SerializeRow` which allows associating given Rust field to a column /
bind marker with given name, instead of matching on the Rust field name.
The motivation is similar as in the case of `SerializeCql`: sometimes,
users might want to use a different naming scheme in Rust and for the
column names.