scylla-macros: add skip_name_checks attribute to SerializeCql
Introduce an attribute to the `SerializeCql` macro which causes the
generated code to skip checking names of the serialized Rust fields
against the UDT field names. The motivation behind the attribute is to
ease transition off the old `IntoUserType` macro which blindly
serialized the fields of the struct as they are defined in the Rust
code. While it disables the name checks, type checking is still done, so
there is protection against type confusion, at least.