scylla-macros: add skip_name_checks attribute to SerializeRow
Introduces an attribute to `SerializeRow` macro which causes name checks
to be skipped when serializing in `enforce_order` flavor. The motivation
is the same as in the case of `SerializeCql` - the old `ValueList` macro
didn't have access to the bind marker names and types, and disabling the
name check might make it easier for some to transition off the old
macro.