macros: Add `force_exact_match` attribute to SerializeCql
This commit changes default behavior of SerializeCql: now it doesn't
return error if there is additional field in UDT, not present in Rust
struct. This is to support adding fields to UDT in database.

New attribute, called `force_exact_match`, is added to opt-in to
previous behavior.

Documentation is adjusted, and failing test fragments are removed.