serialize: add `ValueAdapter` and `ValueListAdapter`
Add newtype wrappers that implement `SerializeRow`/`SerializeCql` if the
type wrapped over implements `ValueList`/`Value`. It should be useful
when migrating codebases to the new traits gradually so that objects
from the unchanged code which only implement the old trait can be passed
to the code which was already updated to the new traits.