scylla-cql: work around the lint about supposedly no-op .map
In some places, we have the following expression:

    x.iter().map(|(k, v)| (k, v))

Although the .map call looks like a no-op, it changes the type of the
argument from `&(K, V)` to `(&K, &V)`. This happens due to the so-called
"reference matching ergonomics" feature of Rust which allows binding a
reference to a tuple to the `(k, v)` expression and allows referring to
the tuple's constituents by reference (`k: &K`, `v: &V`).

The issue is already known and reported and will most likely be fixed
in a future release, but for the time being let's reformulate the code
to avoid the lint like this:

    x.iter().map(|p| (&p.0, &p.1))