scylla-macros: Replace derive macros docs with stubs
When both original item and re-export are documented, documentations are
merged. Before this commit, both original item and re-export were
documented - identically or at least very similarly. This caused docs
in scylla-cql crate to be copy-pasted twice. Replace docs with stubs
pointing to the scylla crate to avoid it and point users to proper
documentation.